+++ packages/ispectify_dio/lib/src/data/error.dart
@@ -1,4 +1,5 @@
 import 'package:dio/dio.dart';
+import 'package:ispectify/ispectify.dart';
 import 'package:ispectify_dio/src/data/_data.dart';
 
 class DioErrorData {
@@ -12,12 +13,29 @@ class DioErrorData {
   final DioRequestData requestData;
   final DioResponseData responseData;
 
-  Map<String, dynamic> get toJson => {
+  Map<String, dynamic> toJson({
+    RedactionService? redactor,
+    Set<String>? ignoredValues,
+    Set<String>? ignoredKeys,
+  }) =>
+      {
         'type': exception?.type,
         'error': exception?.error,
         'stack-trace': exception?.stackTrace,
         'message': exception?.message,
-        'request-options': requestData.toJson,
-        'response': responseData.toJson,
+        'request-options': redactor == null
+            ? requestData.toJson()
+            : requestData.toJson(
+                redactor: redactor,
+                ignoredValues: ignoredValues,
+                ignoredKeys: ignoredKeys,
+              ),
+        'response': redactor == null
+            ? responseData.toJson()
+            : responseData.toJson(
+                redactor: redactor,
+                ignoredValues: ignoredValues,
+                ignoredKeys: ignoredKeys,
+              ),
       };
 }